Created detecting sound and designing official map
Player makes sound. Need to update it to change the sound output and length from where it could be heard whether or not the player is crouching or not. AI can hear sound and will turn towards the sound being heard. Started designing the official map. Getting the layout and a lot of the blueprints for the layout pieces being created and ready to be layed out to create the map.