Created a foot skeletal mesh for the FollowAI  & fixing AI navigation through doors
The FollowAI now has feet as the outline effect rather than 2 planes. The player can now see their footsteps whenever they 're within a certain distance of the AI. AI can't navigate through doors. So i've been creating a DoorTravelPoints class that'll telport the AI through the door frame. Just need to increase the complexity of the code to see whether or not the player is within the room or not and to teleport the AI based on that information.