Fixed door travel points issue
Fixed issue where the AI wouldn't go through door travel points because it would only check on if the AI begun overlapping instead of continuing to check becuase the AI is faster than the player. The Door travel points will continously check if the AI is overlapping with any of the door travel points. Have started the sound changes I need to make for the players footsteps. The sound and distance the sound can be heard from the players foostep is based on if the player is crouching or not crouching.