Created vents
Altered the game to where the AI's hearing threshold becomes super high that they can almost hear the player all around the map and the AI will jump through a vent near them. I want to change this to where only certain sounds and actions will actually trigger it. But for now and testing purposes, i'll leave it the way that it is.