Random patrol point logic & altered AI logic to teleport through vents
Created the FollowAI's ability to go to random patrol points during their normal patrol path. AI had an issue to where they'd still teleport to a vent even when the player isn't in a room. Now it'll make sure that they player is in a room while also checking to see which one is closer, the AI or the vent to the player.