Added more sound effects for combat. Some work in Official map.
Added take damage and hit blocked sound effects for AI and player.  Fixing issue where AI is continuously blocking. Created combatUI so the player knows the health of themand the AI. *May want to remove this. Added keys all around the map and locked doors for player to find. Added collision boxes to keys for a better way to find them.